Refactor
1. Generally speaking you shouldn't manipulate dom properties directly
   like you did with the button.  It also was a separation of concerns
   issue. The App component was managing state of a child component
   directly without the child component knowing about it.  Instead what
   I did was added a new prop called `running` that the child prop could
   use as way to determine its own state and render the button or other
   fields within, appropriately.  It also makes TestPanel as a component
   far more powerful and isolated.

2. Cleaned up runTest APIs
   The APIs for running tests were very tightly coupled to the HTML you
   generate.  This coupling makes it hard to re-use this component or
   this functionality in other places.  What if some sort of higher
   level component wanted to run a test?  Also, grabbing dom elements
   and forcing click on them is also a no-no based on the previous
   point.  Instead I made runTest and runAllTests focus on a singular
   action - run the test based on the id you pass in, or iterate over
   all the tests and call runTest with that tests ID.  Furthermore, I
   separated out the handler functions - essentially wrappers for the
   click events.  This too goes back to the initial point of this bullet
   - APIs having a singular purpose - the handlers are like the glue -
   they handle the DOM click event, but then call the specific runTest
   API as necessary.

One overall big issue that I did not get into with this commit was how
the data was structured.  There is some complexity in terms of using an
array within local state, with a bunch of objects that are
representative of state.  I was trying where possible to ensure
immutability - except in one very specific case - the `tests` array I
was manipulating directly inline. I as treating it as a mutable array
(that I could mutate it).  Generally speaking, in the world of
functional programming and React, you want to err on the side of
immutability.  This means that if you are going to change an object -
duplicate it first, then change the dupe.

With the way this component is built now, treating the array as
immutable would mean you would have to refactor a bunch of your runTests
to utilize a bunch of nested functions.  For now, I left it as is and
mutated the array inline.

If I was tackling this project from scratch, I would probably actually
setup some sort of store system that could much more cleanly handle
immutability instead of depending on the local component state for it.